+++ Ziewaar.RAD.Doodads.FormsValidation/Services/Support/Streaming/Readers/ICountingEnumerator.cs
@@ -2,6 +2,6 @@ namespace Ziewaar.RAD.Doodads.FormsValidation.Services.Support.Streaming.Readers
 public interface ICountingEnumerator<TType> : IEnumerator<TType>
 {
     public bool AtEnd { get; }
-    public int Cursor { get; }
+    public long Cursor { get; }
     public string? ErrorState { get; set; }
 }
